Update CLI to prompt for solution file only when multiple solution files are present
